+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/store/onheapstore/RefDataOnHeapStore.java
@@ -242,6 +242,12 @@ public class RefDataOnHeapStore extends AbstractRefDataStore {
     }
 
     @Override
+    public void purge(final long refStreamId, final long partNo) {
+        throw new UnsupportedOperationException("Purge functionality is not supported for the on-heap store " +
+                "as the data is transient");
+    }
+
+    @Override
     public void logAllContents() {
         logAllContents(LOGGER::debug);
     }